requirements.txt
flask
openai-whisper
pytest
requests
sounddevice
